AuthorizationModule: move grantRole & transferAdminship inside of AuthorizationModule
