+++ contracts/modules/wrapper/mandatory/BurnModule.sol
@@ -4,13 +4,12 @@ pragma solidity ^0.8.17;
 
 import "../../../../openzeppelin-contracts-upgradeable/contracts/token/ERC20/ERC20Upgradeable.sol";
 import "../../../../openzeppelin-contracts-upgradeable/contracts/proxy/utils/Initializable.sol";
-import "../optional/AuthorizationModule.sol";
+import "../../security/AuthorizationModule.sol";
 
 abstract contract BurnModule is ERC20Upgradeable, AuthorizationModule {
-    bytes32 public constant BURNER_ROLE = keccak256("BURNER_ROLE");
     event Burn(address indexed owner, uint256 amount);
 
-    function __BurnModule_init(string memory name_, string memory symbol_) internal onlyInitializing {
+    function __BurnModule_init(string memory name_, string memory symbol_, address admin) internal onlyInitializing {
         /* OpenZeppelin */
         __Context_init_unchained();
         __ERC20_init_unchained(name_, symbol_);
@@ -20,7 +19,7 @@ abstract contract BurnModule is ERC20Upgradeable, AuthorizationModule {
         __AccessControl_init_unchained();
         
         /* Wrapper */
-        __AuthorizationModule_init_unchained();
+        __AuthorizationModule_init_unchained(admin);
 
         /* own function */
         __BurnModule_init_unchained();
