+++ contracts/modules/wrapper/mandatory/EnforcementModule.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.17;
 
 import "../../../../openzeppelin-contracts-upgradeable/contracts/security/PausableUpgradeable.sol";
 import "../../../../openzeppelin-contracts-upgradeable/contracts/proxy/utils/Initializable.sol";
-import "../optional/AuthorizationModule.sol";
+import "../../security/AuthorizationModule.sol";
 import "../../internal/EnforcementModuleInternal.sol";
 
 /**
@@ -15,11 +15,10 @@ import "../../internal/EnforcementModuleInternal.sol";
 abstract contract EnforcementModule is EnforcementModuleInternal,
     AuthorizationModule {
 
-    bytes32 public constant ENFORCER_ROLE = keccak256("ENFORCER_ROLE");
     string internal constant TEXT_TRANSFER_REJECTED_FROZEN =
         "The address is frozen";
 
-    function __EnforcementModule_init() internal onlyInitializing {
+    function __EnforcementModule_init(address admin) internal onlyInitializing {
         /* OpenZeppelin */
         __Context_init_unchained();
         // AccessControlUpgradeable inherits from ERC165Upgradeable
@@ -31,7 +30,7 @@ abstract contract EnforcementModule is EnforcementModuleInternal,
         __Enforcement_init_unchained();
         
         /* Wrapper */
-        __AuthorizationModule_init_unchained();
+        __AuthorizationModule_init_unchained(admin);
 
         /* own function */
         __EnforcementModule_init_unchained();
