+++ contracts/modules/wrapper/mandatory/SnapshotModule.sol
@@ -3,7 +3,7 @@
 pragma solidity ^0.8.17;
 
 import "../../../../openzeppelin-contracts-upgradeable/contracts/proxy/utils/Initializable.sol";
-import "../optional/AuthorizationModule.sol";
+import "../../security/AuthorizationModule.sol";
 import "../../internal/SnapshotModuleInternal.sol";
 
 /**
@@ -12,9 +12,8 @@ import "../../internal/SnapshotModuleInternal.sol";
  * Useful to take a snapshot of token holder balance and total supply at a specific time
  */
 abstract contract SnasphotModule is SnapshotModuleInternal, AuthorizationModule {
-    bytes32 public constant SNAPSHOOTER_ROLE = keccak256("SNAPSHOOTER_ROLE");
 
-    function __SnasphotModule_init(string memory name_, string memory symbol_) internal onlyInitializing {
+    function __SnasphotModule_init(string memory name_, string memory symbol_, address admin) internal onlyInitializing {
         /* OpenZeppelin */
         __Context_init_unchained();
         // SnapshotModuelInternal inherits from ERC20
@@ -28,7 +27,7 @@ abstract contract SnasphotModule is SnapshotModuleInternal, AuthorizationModule 
         __Snapshot_init_unchained();
 
         /* Wrapper */
-        __AuthorizationModule_init_unchained();
+        __AuthorizationModule_init_unchained(admin);
 
         /* own function */
         __SnasphotModule_init_unchained();
