+++ package.json
@@ -25,9 +25,8 @@
     "test:burn": "npx truffle test test/standard/modules/BurnModule.test.js test/proxy/modules/BurnModule.test.js",
     "test:validation": "npx truffle test test/standard/modules/ValidationModule.test.js test/proxy/modules/ValidationModule.test.js",
     "test:enforcement": "npx truffle test test/standard/modules/EnforcementModule.test.js test/proxy/modules/EnforcementModule.test.js",
-    "test:authorization": "npx truffle test test/standard/modules/AuthorizationModule.test.js test/proxy/modules/AuthorizationModule.test.js",
-    "test:base": "npx truffle test test/standard/modules/BaseModule.test.js test/proxy/modules/BaseModule.test.js",
-    "test:transferAdminship": "npx truffle test test/standard/TransferAdminship.test.js test/proxy/TransferAdminship.test.js"
+    "test:authorization": "npx truffle test test/standard/modules/AuthorizationModule/AuthorizationModule.test.js test/proxy/modules/AuthorizationModule/AuthorizationModule.test.js test/standard/modules/AuthorizationModule/TransferAdminship.test.js test/proxy/modules/AuthorizationModule/TransferAdminship.test.js",
+    "test:base": "npx truffle test test/standard/modules/BaseModule.test.js test/proxy/modules/BaseModule.test.js"
   },
   "repository": {
     "type": "git",
