+++ test/proxy/modules/AuthorizationModule/TransferAdminship.test.js
@@ -1,9 +1,9 @@
 const CMTAT = artifacts.require('CMTAT')
 const { deployProxy } = require('@openzeppelin/truffle-upgrades')
-const TransferAdminshipCommon = require('../common/TransferAdminshipCommon')
+const TransferAdminshipCommon = require('../../../common/AuthorizationModule/TransferAdminshipCommon')
 
 contract(
-  'Standard - TransferAdminship',
+  'Proxy - TransferAdminship',
   function ([_, oldAdmin, newAdmin, attacker]) {
     beforeEach(async function () {
       this.cmtat = await deployProxy(CMTAT, [true, oldAdmin, 'CMTA Token', 'CMTAT', 'CMTAT_ISIN', 'https://cmta.ch'], {
